repo_content/nexus_cli/requirements.txt
-e ../nexus_shared
click
rich
google-genai
httpx
pydantic
